Only format openapi.yaml
